not all installations use 0777
you can get the individual setting as string from $config->chmodDir for directories or as string from $config->chmodFile for files. And as PHPs chmod internally convert strings to integer (777 instead of 0777) we also need to use octdec("0777").

A closing PHP tag on the file end is not really necessary. You can use it, but you don't need to.

Nice Module! :)